+++ src/framework/core/code-contrib/get.ts
@@ -12,14 +12,15 @@ import { frameworkStore } from "@/core/store";
  * an empty map is returned.
  *
  * @category Extension Contribution API
+ * @typeParam Value - Expected type of the code contribution value
  * @param contribPointId - The code contribution point identifier.
  * @returns A read-only map of code contributions
  *  or `undefined` if it does not exist.
  */
-export function getCodeContributions<Data>(
+export function getCodeContributions<Value>(
   contribPointId: string
-): ReadonlyMap<string, Data> | undefined {
+): ReadonlyMap<string, Value> | undefined {
   return frameworkStore.getState().codeContributions[
     contribPointId
-  ] as ReadonlyMap<string, Data>;
+  ] as ReadonlyMap<string, Value>;
 }